requirements.txt
torch==2.0.0
torchvision==0.15.1
albumentations==1.3.0
opencv-python==4.6.0.66
imageio==2.9.0
numpy==1.23.1
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit==1.12.1
einops==0.3.0
transformers==4.19.2
webdataset==0.2.5
kornia==0.6
open_clip_torch==2.0.2
invisible-watermark>=0.1.5
streamlit-drawable-canvas==0.8.0
torchmetrics==0.6.0
xformers==0.0.16
triton
matplotlib
wandb
pillow
waveprop==0.0.10
sacred
recordclass
-e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
-e git+https://github.com/openai/CLIP.git@main#egg=clip
# -e .
